Refactor campaign start logic and update genome methods
Refactored the `Start` method in `ICampaign.cs` to change the default
value of `startCycle` from `-1` to `0`. In `Campaign.cs`, moved
`StartSeeded` method above `Start` and modified its implementation.
Updated `Start` method to check if `startCycle` is not `0` and added
comments for clarity. Renamed `adamGenome` to `basicGenome` in
`InitializeGenomePool`.

In `GeneticOperator.cs`, added a commented-out line in `Mutate` to
retrieve genes. Updated `Genome.cs` to call `GeneticOperator.Mutate`
in `Mutate` method and added a commented-out line for the previous
implementation. Added a line in `Speciate` method to return sibling
genomes.